Drop inconsistent begin/end hook from Data/Trigger
When the robot is disabled, the run() method will be called after the
last end() is called (during the last teleop/auton period), and before a
call is made to before() (during the next teleop/auton period). This
could lead to unexpected errors.

Resolving this inconsistency would introduce more complexity than the
feature is worth. Dropping these hooks has a cascading bonus effect of
reducing consistency across a number of interrelated classes.