Record log output to disk
The major shortcoming, due to the limitations of the Java API we have to
work with, is the lack of Exception#getStackTrace(). There's really no
way to get a stack trace into a String in Java, or to write one to a
file. As a result, we try to do the next-best-ish thing and record the
output of Exception#toString().

In addition to this logging, I'd suggest leaving a NetConsole window
open on your driver station computer during matches. Set the buffer to a
really long value, and use that to record output generated during the
match. Make sure your drivers know not to turn off or log out of the
driver station after a match!