Attempt to guard against random exceptions
At the expense of additional complexity, the framework now attempts to
catch exceptions arising in various points of the control loop: during
data and trigger updates, during the three steps of action execution,
during connector conduction, and during each overall loop tick. When an
exception is intercepted, it is logged, and execution tries to continue
onward anyway. This can be invaluable in a match situation, where one
NullPointerException somewhere in the stack could otherwise stop your
robot dead in its tracks for the rest of the match.

Of course, this extra safety can be disabled, by making a call to
Safety#disable(). A nice big warning is logged when the code runs like
this.

During testing, it is actually strongly recommended that you disable
Safety, so that problems can be caught while you're working on the
robot. Just don't forget to turn it back on before going out to a match,
where the robot's ability to move around is more important than anything
else.